// src/app/authentication/components/search-responsibles/search-responsibles.interface.ts
export interface ISearchresponsible {
    address: string;
    id_work: string;
    r_villagecode: string;
    r_villagecodefull: string;
    firstname: string;
    lastname: string;
    position: string;
    hosname: string;
    villagename: string;
    tambonname: string;
    amphurname: string;
    changwatname: string;
    work: string;
}